requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msal==1.35.0
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.6
